refactor tests for invitations plugin
